Add compile-if macro and optimize replace-content
To support legacy versions of clojure, compile-if was added which evaluates if
condition at compile time and inject the then or else form as required. Used to
support 1.2 and it's lack of IRecord as well as unsafe versions where into lost
meta data.

Optimized replace-content to do less work if the old and new values were the
same type.