Add <> symbol to nearly all synthread macros
The diamond (<>) symbol is now bound and rebound to the current topic as it is
threaded through the forms. This behavior is consistent across all macros
except in the as and aside macros.

Actually those two macros may no longer be needed and so deprecated.
;; equivalent
(->/as foo ...)
(->/let [foo <>] ...)

;; equivalent
(->/aside foo ...)
(->/let [foo <>] (doto (do ...)))