Initial refactoring of synthread.cljc
There are two intertwined refactorings:
  1. 'cloaking' vars
  2. cljc compatible defmacro

cloaking vars provide the ability to defer the publishing of a cloaked function
or macro until a later point in the file. Which allows those definitions to be
written conveniently based on the original definitions.

The cljc compatible defmacro is designed for use in cljc files to look the same
in both clj and cljc contexts without using #? reader literals.

The decloak macro is also cljc compatible doing the right thing in
clj and cljs contexts.